fix non page files
