--- src/pages/catalog/api.ts
@@ -1,5 +1,0 @@
-import { ListItem } from '@types/catalog';
-
-export function fetchItems(): Promise<ListItem[]> {
-  return fetch('http://localhost:3000//api/items').then((res) => res.json());
-}
