+++ src/pages/login/PasswordForm.tsx
@@ -1,6 +1,5 @@
 import { Button, PasswordInput } from '@mantine/core';
 import { useForm } from '@mantine/form';
-import { login } from './api';
 
 interface PasswordFormValues {
   password: string;
@@ -41,4 +40,11 @@ export default PasswordForm;
 
 function canNotBeEmpty(value: string) {
   return !value && 'Can not be empty';
+}
+
+function login(username: string, password: string) {
+  return fetch('/api/login', {
+    method: 'post',
+    body: JSON.stringify({ username, password })
+  });
 }
